+++ .gitignore
@@ -1,0 +1,3 @@
+
+BT/Assets/StreamingAssets/LevelGen/~$LevelGenDesigner.xlsm
+BT/Assets/StreamingAssets/LevelGen/~$LevelGenDesigner.xlsm.meta
