+++ BT/Assets/Prefabs/Unity Store Assets/Meshab prefabs/medical props/infusion stand.prefab
@@ -23,13 +23,13 @@ Transform:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 3005477498573440269}
-  m_LocalRotation: {x: 0.00000002185566, y: 1, z: -0.000000021855664, w: 0.0000014603136}
+  m_LocalRotation: {x: 0, y: 0.7071068, z: 0, w: 0.7071068}
   m_LocalPosition: {x: 0, y: 1.61, z: 0.139}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 6491855605825271534}
   m_RootOrder: 0
-  m_LocalEulerAnglesHint: {x: 0, y: 180.00002, z: 0}
+  m_LocalEulerAnglesHint: {x: 0, y: 90, z: 0}
 --- !u!1 &6491855605824776910
 GameObject:
   m_ObjectHideFlags: 0
@@ -42,7 +42,7 @@ GameObject:
   - component: {fileID: 6491855605826041550}
   - component: {fileID: 6491855605827040270}
   - component: {fileID: 6491855605822669934}
-  - component: {fileID: 3304003165735176541}
+  - component: {fileID: 4563696486851256449}
   m_Layer: 0
   m_Name: infusion stand
   m_TagString: Untagged
@@ -126,7 +126,7 @@ MeshCollider:
   m_Convex: 0
   m_CookingOptions: 30
   m_Mesh: {fileID: 4300000, guid: 28ca43ebe41e9c845be2dadd79a60a48, type: 3}
---- !u!114 &3304003165735176541
+--- !u!114 &4563696486851256449
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -135,8 +135,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 6491855605824776910}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: c532e4c9dd047a44a8e72e9c9844a897, type: 3}
+  m_Script: {fileID: 11500000, guid: 621fbfd6046f9824b9ec093a09487645, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  infusionBagLocs:
-  - {fileID: 4546871925505878374}
+  subObjectLocs: []
