+++ BT/Assets/Scripts/LevelGen/SubObjectParent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 621fbfd6046f9824b9ec093a09487645
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
