+++ BT/Assets/Scripts/LevelGen/SubObjectParentFixed.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class SubObjectParentFixed : SubObjectParent
+{
+    void Awake()
+    {
+        // Asset creator must set all transforms that are used as SubObjectLocs in the Unity editor
+    }
+}
