+++ BT/Assets/Scripts/LevelGen/SubObjectParentFixed.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c532e4c9dd047a44a8e72e9c9844a897
+guid: 7c0407ec90c9f3e4e989febf4835b385
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
