+++ BT/Assets/Scripts/UITextShower.cs
@@ -15,8 +15,7 @@ public class UITextShower : MonoBehaviour
 
     void Update()
     {
-        //scoreText.text = "Total Points " + scoreTracker.GetComponent<ObjectMessageHandler>().pointTotal + "/9";
-        // Putting this here until somebody adds the pointTotal back to the ObjectMessageHandler
-        scoreText.text = "Total Points ERROR/9";
+        // This should be rewritten to only update the point display when the player's score changes, not on Update()
+        scoreText.text = "Total Points " + scoreTracker.GetComponent<ObjectMessageHandler>().pointTotal + "/9";
     }
 }
