--- BT/Assets/StreamingAssets/04.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7e4e556b56ec05d41a1d13451669473e
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
