--- BT/Assets/StreamingAssets/Documentation.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8c5c8d696f76eb848abb873e81b43e64
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
