--- BT/Assets/StreamingAssets/Fonts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 1271ffadeffeec947b799894c6870c31
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
