+++ BT/Assets/StreamingAssets/LevelGen/LevelGenTestExport.txt
@@ -1,57 +1,57 @@
-Room	0	RoomXLGPatient_base			
-Bed	1	ExaminationBed01			
-Wall	1				
-BedsideHead	1				
-BedsideMid	1				
-BedsideFoot	1				
-Cart	1				
-Screen	1				
-Screen	2				
-Bed	2	ExaminationBed01			
-Wall	2				
-BedsideHead	2				
-BedsideMid	2				
-BedsideFoot	2				
-Cart	2				
-Screen	3				
-Screen	4				
-Bed	3	ExaminationBed01			
-Wall	3				
-BedsideHead	3				
-BedsideMid	3				
-BedsideFoot	3				
-Cart	3				
-Screen	5				
-Screen	6				
-Bed	4	ExaminationBed01			
-Wall	4				
-BedsideHead	4				
-BedsideMid	4				
-BedsideFoot	4				
-Cart	4				
-Screen	7				
-Screen	8				
-Bed	5	ExaminationBed01			
-Wall	5	XRayViewBox01			
-BedsideHead	5	RackInfusionBag01	InfusionBag	1	InfusionBag02
-BedsideMid	5				
-BedsideFoot	5				
-Cart	5	Trolley01			
-Screen	9				
-Screen	10				
-Bed	6	ExaminationBed01			
-Wall	6				
-BedsideHead	6				
-BedsideMid	6				
-BedsideFoot	6				
-Cart	6				
-Screen	11	PrivacyScreen01			
-Screen	12	PrivacyScreen01			
-					
-					
-					
-					
-					
-					
-					
-					
+Room	0	RoomXLGPatient_base												
+SuzieWalk	1	ExaminationBed01												
+Wall	1													
+BedsideHead	1													
+BedsideMid	1													
+BedsideFoot	1													
+Cart	1													
+Screen	1													
+Screen	2													
+Bed	2													
+Wall	2													
+BedsideHead	2	RackInfusionBottle01	InfusionBottle	1	InfusionBottle01	InfusionBottleLiquid	1	InfusionBottleLiquid01	InfusionBottle	2	InfusionBottle02	InfusionBottleLiquid	2	InfusionBottleLiquid01
+BedsideMid	2	RackInfusionBottle01	InfusionBottle	1	InfusionBottle02	InfusionBottleLiquid	1	InfusionBottleLiquid02	InfusionBottle	2	InfusionBottle04	InfusionBottleLiquid	2	InfusionBottleLiquid03
+BedsideFoot	2	RackInfusionBottle01	InfusionBottle	1		InfusionBottleLiquid	1		InfusionBottle	2	InfusionBottle01	InfusionBottleLiquid	2	InfusionBottleLiquid02
+Cart	2													
+Screen	3													
+Screen	4													
+Bed	3													
+Wall	3													
+BedsideHead	3													
+BedsideMid	3													
+BedsideFoot	3	RackInfusionBottle01	InfusionBottle	1	InfusionBottle04	InfusionBottleLiquid	1		InfusionBottle	2	InfusionBottle01	InfusionBottleLiquid	2	InfusionBottleLiquid02
+Cart	3													
+Screen	5													
+Screen	6													
+Bed	4													
+Wall	4													
+BedsideHead	4													
+BedsideMid	4													
+BedsideFoot	4													
+Cart	4													
+Screen	7													
+Screen	8													
+Bed	5													
+Wall	5													
+BedsideHead	5													
+BedsideMid	5													
+BedsideFoot	5													
+Cart	5													
+Screen	9													
+Screen	10													
+Bed	6													
+Wall	6													
+BedsideHead	6													
+BedsideMid	6													
+BedsideFoot	6													
+Cart	6													
+Screen	11													
+Screen	12													
+														
+														
+														
+														
+														
+														
+														
+
