+++ BT/Assets/StreamingAssets/LevelGen/~$LevelGenDesigner.xlsm.meta
@@ -1,6 +1,5 @@
 fileFormatVersion: 2
-guid: fdcf13ee8b63e5945b1a4784329267e2
-folderAsset: yes
+guid: d0983897663ccb446bfb407068c0e9db
 DefaultImporter:
   externalObjects: {}
   userData:
