--- BT/Assets/StreamingAssets/Resources/Fonts & Materials.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 950937f1437384a499d18f9529b90b4e
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
