--- BT/Assets/StreamingAssets/Resources/Sprite Assets.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 264fff8175a0447439dd7780ceba54c9
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
