--- BT/Assets/StreamingAssets/Resources/Style Sheets.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b3ee7b9caf7b68b4c9e6d21eb0267b68
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
