--- BT/Assets/StreamingAssets/Sprites.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: aecf35bf61a584f4b8ee5d3532f4f021
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
